// React/app/GerAdv_TS/AgendaSemana/Interfaces/interface.AgendaSemana.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IAgendaSemana {
// 202501251
    id: number;
 
	funcionario: number,
	advogado: number,
	tipocompromisso: number,
	cliente: number,
	paranome: string,
	data: string,
	hora: string,
	compromisso: string,
	concluido: boolean,
	liberado: boolean,
	importante: boolean,
	horafinal: string,
	nome: string,
	nomecliente: string,
	tipo: string,
}

export interface IAgendaSemanaFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IAgendaSemanaIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}